Revert branches

+++ common/src/main/scala/models/TopicCount.scala
@@ -2,7 +2,7 @@ package models
 
 import play.api.libs.json.{Format, Json}
 
-case class TopicCount(topicName: String,  registrationCount: Int)
+case class TopicCount(topicName: String,  registrationCount: Long)
 
 object TopicCount{
   implicit val topicCountJF: Format[TopicCount] = Json.format[TopicCount]
